requirements.txt
Django==3.2.8
django-crispy-forms==1.12.0
django-rest-framework==3.12.4
Pillow==8.4.0
